Update: script now asks for which links to make

+++ setup/set-links.sh
@@ -4,12 +4,15 @@ set -e
 
 echo 'linking config from .dotfiles to system .config'
 echo '[WARN] watch out! this script will remove your .config files'
-read -p "Want to continue and link files: [y/n] " ANSWER
+read -p "bashrc link: [y/n] " ANSWER
 if [[ $ANSWER == "y" ]]; then
 	rm -rf ~/.bashrc
 	ln -s ~/.dotfiles/config/bashrc.config ~/.bashrc
 	echo 'added link for .bashrc'
+fi
 
+read -p "tmux link: [y/n] " ANSWER
+if [[ $ANSWER == "y" ]]; then
 	# checks if dir does not exists
 	if [[ ! -d ~/.tmux/plugins/tpm ]]; then
 		git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
@@ -17,7 +20,10 @@ if [[ $ANSWER == "y" ]]; then
 	rm -rf ~/.tmux.conf
 	ln -s ~/.dotfiles/config/tmux.config ~/.tmux.conf
 	echo 'added link for tmux'
+fi
 	
+read -p "i3 link: [y/n] " ANSWER
+if [[ $ANSWER == "y" ]]; then
 	rm -rf ~/.config/i3
 	mkdir ~/.config/i3
 	ln -s ~/.dotfiles/config/i3.config ~/.config/i3/config
@@ -32,18 +38,27 @@ if [[ $ANSWER == "y" ]]; then
 	mkdir ~/.config/picom
 	ln -s ~/.dotfiles/config/picom.config ~/.config/picom/config
 	echo 'added link for picom'
+fi
 	
+read -p "nvim link: [y/n] " ANSWER
+if [[ $ANSWER == "y" ]]; then
 	rm -rf ~/.config/nvim
 	ln -s ~/.dotfiles/nvim ~/.config/nvim
 	echo 'added link for nvim'
+fi
 	
+read -p "kitty link: [y/n] " ANSWER
+if [[ $ANSWER == "y" ]]; then
 	rm -rf ~/.config/kitty
 	mkdir ~/.config/kitty
 	ln -s ~/.dotfiles/config/kitty.config ~/.config/kitty/kitty.conf
 	ln -s ~/.dotfiles/config/kitty-launch.config ~/.config/kitty/kitty-launch.config
 	ln -s ~/.dotfiles/config/kitty-colorscheme.config ~/.config/kitty/kitty-colorscheme.config
 	echo 'added links for kitty'
+fi
 
+read -p "wezterm link: [y/n] " ANSWER
+if [[ $ANSWER == "y" ]]; then
 	rm -rf ~/.config/wezterm
 	mkdir ~/.config/wezterm
 	ln -s ~/.dotfiles/config/wezterm.lua ~/.config/wezterm/wezterm.lua
